Extract limited deque out of email sender
