+++ src/main/kotlin/com/appifyhub/monolith/service/integrations/email/LogEmailSender.kt
@@ -1,10 +1,9 @@
 package com.appifyhub.monolith.service.integrations.email
 
 import com.appifyhub.monolith.domain.creator.Project
+import com.appifyhub.monolith.service.integrations.limitedDeque
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
-import kotlin.properties.ReadWriteProperty
-import kotlin.reflect.KProperty
 
 @Component
 class LogEmailSender : EmailSender {
@@ -35,36 +34,3 @@ class LogEmailSender : EmailSender {
   }
 
 }
-
-/**
- * Produces a delegated property of ArrayDeque type.
- * The custom property holds the collection internally and limits it to a maximum of 'limit' items.
- * Limiting/trimming happens on each read and write of the property.
- *
- * @param limit How big can the collection grow before it gets trimmed
- * @return Kotlin's delegated property
- * @sample LogEmailSender.history
- */
-@Suppress("SameParameterValue")
-private fun <E> limitedDeque(limit: Int): ReadWriteProperty<Any, ArrayDeque<E>> =
-  object : ReadWriteProperty<Any, ArrayDeque<E>> {
-
-    private var deque: ArrayDeque<E> = ArrayDeque(limit)
-
-    override fun getValue(thisRef: Any, property: KProperty<*>): ArrayDeque<E> {
-      applyLimit()
-      return deque
-    }
-
-    override fun setValue(thisRef: Any, property: KProperty<*>, value: ArrayDeque<E>) {
-      deque = value
-      applyLimit()
-    }
-
-    private fun applyLimit() {
-      while (deque.size > limit) {
-        deque.removeFirst()
-      }
-    }
-
-  }
